+++ resources/js/dashboard/adminDashboard/AdminCategoriesList.vue
@@ -36,13 +36,30 @@ export default{
         }
     },
     methods:{
-    deleteCategory(id){
-        axios.delete('/api/categories/' + id)
-        .then((response)=>{ 
-            console.log(response);
-            this.fetchCategories(); // Fetch the updated list after deletion
-        }).catch((error)=>{
-            console.log(error);
+      deleteCategory(id) {
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            axios.delete('/api/categories/' + id)
+              .then((response) => { 
+                console.log(response);
+                this.fetchCategories(); // Fetch the updated list after deletion
+                Swal.fire({
+                  title: "Deleted!",
+                  text: "Your file has been deleted.",
+                  icon: "success"
+                });
+              }).catch((error) => {
+                console.log(error);
+              });
+          }
         });
       },
     fetchCategories() {
